opt: add hieracrhical region analysis, lack exit-merge
next step, implement the step 5: exit merge.
I need to add a variable for each exit, and exit through the same
region.
One thing I'm not sure about is: we do exit the region, but how to
handle loop exit? Shall I branch back to the loop header, or not?

Signed-off-by: Nathan Gauër <brioche@google.com>